\name{plot-methods}
\docType{methods}
\alias{plot-methods}
\alias{plot,gExperimentSetup,missing-method}
\title{ Methods for Function \code{plot}}
\description{
 Create bar plots based on the experiment setup.
}
\section{Methods}{
\describe{

\item{\code{signature(x = "gExperimentSetup", y = "missing")}}{
Plot distribution of relevant variables over different batches.
}
}}
\keyword{methods}
%\keyword{ ~~ other possible keyword(s) ~~ }
