\name{predefined}
\alias{predefined}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Show predefined objects in the package.
}
\description{
Show predefined objects in the package.
}
\usage{
predefined()
}
%- maybe also 'usage' for other objects documented here.
% \arguments{
% \item{\dots}{ Additional parameters passed to optimization function.
% }
% }
\details{
 Currently layout of two chips, three IlluminaBeadChip plates, and the
 MSA4 robotic loader are defined.
}
%\value{
%%  ~Describe the value returned
%%  If it is a LIST, use
%%  \item{comp1 }{Description of 'comp1'}
%%  \item{comp2 }{Description of 'comp2'}
%% ...
%}
% \references{
% %% ~put references to the literature/web site here ~
% }
% \author{
% %%  ~~who you are~~
% }
% \note{
% %%  ~~further notes~~
% }

%% ~Make other sections like Warning with \section{Warning }{....} ~

% \seealso{

% }
